Add support for Wear OS 4.0 (One UI 5.0)
Samsung now opens the Samsung Pay app using a broadcast from the
PhoneWindowManagerExt class in /system/framework/services.jar.

NOTE: In order to be able to launch an app from the background, this app now
requires the SYSTEM_ALERT_WINDOW permission to bypass restrictions imposed
in Android 10 and later versions.